FinanceApp/stockProject - main/stockProject/monitor/requirements.txt
psycopg2-binary
docker